+++ src/web/src/components/index.ts
@@ -1,5 +1,5 @@
 // Atoms
-export { Button } from "./atoms/Button";
+export { Button } from "@4lt7ab/ui/ui";
 export { Input } from "@4lt7ab/ui/ui";
 export { Select } from "@4lt7ab/ui/ui";
 export { Textarea } from "@4lt7ab/ui/ui";